uxai-ugc-agent/requirements.txt
openai>=1.0.0
requests
praw
flask
flask-socketio
flask-cors
elevenlabs
moviepy==1.0.3
pillow
psutil
python-dotenv
schedule
httpx
gevent
